+++ src/ScrumBoardItBundle/Api/AbstractApi.php
@@ -1,5 +1,5 @@
 <?php
-namespace ScrumBoardItBundle\API;
+namespace ScrumBoardItBundle\Api;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
@@ -55,10 +55,12 @@ abstract class AbstractApi
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($ch, CURLOPT_TIMEOUT, 5);
         $content = curl_exec($ch);
-        $httpCode = curl_getinfo($ch,CURLINFO_HTTP_CODE);
+        $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         curl_close($ch);
         
-        if($httpCode !== 200) throw new HttpException($httpCode);
+        if ($httpCode !== 200)
+            throw new HttpException($httpCode);
+        
         return json_decode($content);
     }
